Update Solution options text (#374)
* Update Solution options text

In some circumstances, exiting and re-entering edit mode will not correct the block on editing. In this scenario, other options such as copying modules to another page or Duplicating a page may also not function as workarounds. While the records can be cleared within the database, a method of workaround within the DNN interface is as follows:

* Spelling correction

Changed "teh" to "the".

* Correcting casing

Changed "Duplicating" to "duplicating"